backend/requirements.txt
python-dotenv
google-api-python-client
google-generativeai
requests
beautifulsoup4
streamlit
scrapegraphai